Added tests for add operations. Fixed add operations.

+++ src/core/formats/DFormat.cpp
@@ -5,7 +5,7 @@ DFormat::DFormat(uint32_t instruction)
     : Format(instruction),
       m_D((instruction >> 21) & 0x1F),
       m_A((instruction >> 16) & 0x1F),
-      m_imm(instruction & 0xFFFF) {}
+      m_imm(static_cast<uint16_t>(instruction & 0xFFFF)) {}
 
 DFormat::~DFormat() {}
 
@@ -13,5 +13,5 @@ uint32_t DFormat::GetD() const { return m_D; }
 
 uint32_t DFormat::GetA() const { return m_A; }
 
-uint32_t DFormat::GetImmediate() const { return m_imm; }
+uint16_t DFormat::GetImmediate() const { return m_imm; }
 }  // namespace PurpleBox
