+++ src/core/formats/DFormat.hpp
@@ -11,12 +11,12 @@ class DFormat : public Format {
 
   uint32_t GetD() const;
   uint32_t GetA() const;
-  uint32_t GetImmediate() const;
+  uint16_t GetImmediate() const;
 
  private:
   uint32_t m_D;
   uint32_t m_A;
-  uint32_t m_imm;
+  uint16_t m_imm;
 };
 }  // namespace PurpleBox
